requirements.txt
datasets
tqdm
Pillow
huggingface_hub
transformers
torch
accelerate
evaluate
numpy
accuracy
codecarbon

